Chapter 2 Challenge
Challenge is to layer plots for a single variable using the variable Petal.Length from "iris" dataset from R.  

Create three different charts, one on top of another, for that variable
First, create a histogram of Petal.Length.  Then, on top of that, superimpose a kernel density plot.  Then, underneath the both of them, put a rug or a line plot